requirements.txt
pandas
pydantic
python-dotenv
tqdm
faker
litellm
requests
